+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 Notable changes to this project are documented here.
 
+## next
+
+Internally refactored code to use [gruber](https://github.com/robb-j/gruber).
+This project was part of the inspiration for gruber so it's fitting to be
+migrated to it, it mostly involved removing code too! 🎉
+
+- **breaking** — by default the server won't commit changes, set `GIT_COMMIT=1`
+  or the `git.commit` configuration to `true`.
+- added support for `.ini` file parsing
+- configure the port to run on with the `--port` flag or `APP_PORT` variable
+- override the result of `GET /` by setting `APP_NAME`, `APP_VERSION` or
+  `app.name` and `app.version` via configuration.
+- See the current configuration and print usage by running the config.ts script
+  ~ `deno run -A source/config.ts`
+  - new configuration `git.commitPrefix` to customise the commit message
+
 ## 0.2.1
 
 Tweak `/changed` to show all changes and not remove duplicates